separating S3 setup
