+++ uploadservice-s3/src/main/java/net/gotev/uploadservice/s3/S3ClientWrapper.kt
@@ -1,17 +1,10 @@
 package net.gotev.uploadservice.s3
 
-import android.app.PendingIntent
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.app.Notification
 import android.content.Context
-import android.content.Intent
-import android.os.Build
 import com.amazonaws.auth.CognitoCachingCredentialsProvider
 import com.amazonaws.mobileconnectors.s3.transferutility.TransferState
 import com.amazonaws.mobileconnectors.s3.transferutility.TransferObserver
 import com.amazonaws.mobileconnectors.s3.transferutility.TransferUtility
-import com.amazonaws.mobileconnectors.s3.transferutility.TransferService
 import com.amazonaws.mobileconnectors.s3.transferutility.TransferListener
 import com.amazonaws.regions.Region
 import com.amazonaws.regions.Regions
@@ -22,7 +15,6 @@ import net.gotev.uploadservice.logger.UploadServiceLogger
 import java.io.Closeable
 import java.io.File
 import java.lang.Exception
-import java.util.UUID
 
 class S3ClientWrapper(
     private val uploadId: String,
@@ -37,35 +29,6 @@ class S3ClientWrapper(
     private val transferUtility = TransferUtility.builder().s3Client(s3).context(context).build()
     private lateinit var transferObserver: TransferObserver
     private lateinit var uploadingFile: UploadFile
-
-    init {
-        val tsIntent = Intent(context, TransferService::class.java)
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val id = UUID.randomUUID().toString()
-            val pendingIntent = PendingIntent.getActivity(context, 0, tsIntent, 0)
-
-            // Notification Manager to listen to a channel
-            val channel = NotificationChannel(id, "Transfer Service Channel", NotificationManager.IMPORTANCE_DEFAULT)
-            val notificationManager = context.getSystemService(NotificationManager::class.java)
-            notificationManager.createNotificationChannel(channel)
-
-            // Valid notification object required
-            val notification = Notification.Builder(context, id)
-                .setContentTitle("Transfer Service Notification")
-                .setContentText("Transfer Service is running")
-                .setContentIntent(pendingIntent)
-                .build()
-
-            tsIntent.putExtra(TransferService.INTENT_KEY_NOTIFICATION, notification)
-            tsIntent.putExtra(TransferService.INTENT_KEY_NOTIFICATION_ID, 15)
-            tsIntent.putExtra(TransferService.INTENT_KEY_REMOVE_NOTIFICATION, true)
-
-            // Foreground service required starting from Android Oreo
-            context.startForegroundService(tsIntent)
-        } else {
-            context.startService(tsIntent)
-        }
-    }
 
     interface Observer {
         fun onStateChanged(client: S3ClientWrapper, uploadFile: UploadFile, id: Int, state: TransferState?)
